Extract Position Cards into separate component Add Show More Accordion Component Add condition to display show more accordion when there are more than 4 card entries in position
